Rename workload queue module
