--- agent/src/workload_queue/mod.rs
@@ -1,19 +1,0 @@
-// Copyright (c) 2023 Elektrobit Automotive GmbH
-//
-// This program and the accompanying materials are made available under the
-// terms of the Apache License, Version 2.0 which is available at
-// https://www.apache.org/licenses/LICENSE-2.0.
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-// License for the specific language governing permissions and limitations
-// under the License.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod workload_command_channel;
-pub mod workload_command_queue;
-
-pub use workload_command_channel::WorkloadCommandChannel;
-pub use workload_command_queue::WorkloadCommandQueue;
